[branch esji]
1. gitignore python project 관련 값 전부 적용
2. 테스트 템플릿 통합
